+++ plugins/modules/hmc_update_upgrade.py
@@ -70,7 +70,7 @@ options:
                       C(ibmwebsite) for update through IBM fixcentral website.
                     - When the location type is set to C(disk), first it looks for the C(build_file) in HMC hard disk
                       if it doesn't exist then it looks for C(build_file) in the Ansible Controller node.
-                    - ibmwebsite location type supports only update operation and HMC1030 releas onwards
+                    - ibmwebsite location type supports only update operation and V10 R2 M1030 release onwards
                 type: str
                 required: true
                 choices: ['disk', 'ftp', 'sftp', 'nfs', 'ibmwebsite']
@@ -110,13 +110,16 @@ options:
                       image is kept.
                       During update of the HMC if I(location_type=disk) and ISO image is kept in Ansible controller node or HMC hard disk,
                       this option should be provided with the ansible control node path in which ISO file or network install image is kept.
+                      If the path specified contains the ISO file name then that specified ISO file will considered for updation.
+                      If the path specified doesnot contain the ISO file name then the specified folder will be searched for ISO files,
+                      sorted in alphabetical order and the 1st ISO will be considered for updation.
                 type: str
             ptf:
                 description:
                     - The name of the PTF to install.
                       This option is required when the ISO image is located on the IBM Fix Central website. Otherwise, this option is not valid.
                       This option is required only when the location_type is 'ibmwebsite'.
-                      This option is available for HMC versions from 1030 onwards.
+                      This option is available for HMC versions from V10 R2 M1030 onwards.
                 type: str
     state:
         description:
@@ -129,7 +132,7 @@ options:
     action:
         description:
             - C(listptf) lists available Hardware Management Console (HMC) updates from the IBM Fix Central website.
-            - This option is available for HMC versions from 1030 onwards
+            - This option is available for HMC versions from V10 R2 M1030 onwards
         type: str
         choices: ['listptf']
 '''
@@ -316,6 +319,12 @@ def image_copy_from_local_to_hmc(module, params):
     hmc_user = params['hmc_auth']['username']
     password = params['hmc_auth']['password']
     local_path = params['build_config']['build_file']
+    iso_name = ""
+    iso_count = 0
+
+    if '.iso' in local_path:
+        local_path = local_path[:local_path.rfind("/")]
+        iso_name = params['build_config']['build_file'].split("/")[-1]
 
     if params['state'] == 'upgraded':
         imageFilesUpg = ['base.img', 'disk1.img', 'hmcnetworkfiles.sum', 'img2a', 'img3a']
@@ -329,6 +338,19 @@ def image_copy_from_local_to_hmc(module, params):
             logger.debug(err)
             raise ParameterError("not able to list files on mentioned local path")
 
+    if params['state'] == 'updated':
+        list_cmd = "ls " + local_path
+        rc, out, err = module.run_command(list_cmd)
+        if rc == 0:
+            files = out.split()
+            if params['state'] == 'updated' and iso_name == "":
+                iso_count = sum(file.endswith(".iso") for file in files)
+            if iso_count > 1:
+                raise Error("There are multiple iso files in specified folder. Please specify the iso file also in build_file")
+        else:
+            logger.debug(err)
+            raise ParameterError("not able to list files on mentioned local path")
+
     mkdir_cmd = 'sshpass -p {0} ssh {1}@{2} mkdir -p network_install'.format(password, hmc_user, hmc_host)
     rc1, out1, err1 = module.run_command(mkdir_cmd)
     if rc1 == 1:
@@ -336,6 +358,8 @@ def image_copy_from_local_to_hmc(module, params):
         raise Error("creation of temporary install directory inside HMC failed")
 
     scp_cmd = 'sshpass -p {0} scp -r {1}/* {2}@{3}:/home/{2}/network_install'.format(password, local_path, hmc_user, hmc_host)
+    if params['state'] == 'updated' and iso_name != "":
+        scp_cmd = 'sshpass -p {0} scp  {1}/{2} {3}@{4}:/home/{3}/network_install'.format(password, local_path, iso_name, hmc_user, hmc_host)
     rc2, out2, err2 = module.run_command(scp_cmd, use_unsafe_shell=True)
     if rc2 == 1:
         logger.debug(err2)
@@ -349,8 +373,9 @@ def image_copy_from_local_to_hmc(module, params):
             if not all(True if each in out3 else False for each in imageFilesUpg):
                 remove_image_from_hmc(module, params)
                 raise Error("copy of image to hmc is incomplete. Necessary files are missing")
+        elif params['state'] == 'updated':
+            return out3
         else:
-            files = out3.split()
             iso_file = None
             for fl in files:
                 if '.iso' in fl:
@@ -397,7 +422,7 @@ def list_ptf(module, params):
     initial_version_details = hmc.listHMCVersion()
 
     if int(initial_version_details["SERVICEPACK"]) < 1030:
-        raise VersionError("List ptf is supported from 1030 version onwards.")
+        raise VersionError("List ptf is supported from V10 R2 M1030 version onwards.")
     else:
         ptf_details = hmc.listHMCPTF('ibmwebsite')
 
@@ -501,7 +526,7 @@ def update_hmc(module, params):
     warning_msg = None
     isBootedUp = False
     is_img_in_hmc = False
-
+    iso_file = None
     if not params['build_config']:
         raise ParameterError("missing options on build_config")
 
@@ -514,24 +539,33 @@ def update_hmc(module, params):
     locationType = params['build_config']['location_type']
 
     if locationType == 'disk':
-        is_img_in_hmc = check_image_in_hmc(module, params)
-        if not is_img_in_hmc:
-            iso_file = image_copy_from_local_to_hmc(module, params)
+        if '.iso' in params['build_config']['build_file']:
+            iso_file = params['build_config']['build_file'].split("/")[-1]
+            is_img_in_hmc = check_image_in_hmc(module, params)
+            if not is_img_in_hmc:
+                iso_file = image_copy_from_local_to_hmc(module, params)
         else:
-            hmc_ls_cmd = "sshpass -p {0} ssh {1}@{2} ls {3}".format(password, hmc_user, hmc_host, params['build_config']['build_file'])
-            rc, out, err = module.run_command(hmc_ls_cmd)
-            if rc == 0:
-                files = out.split()
-                for fl in files:
-                    if '.iso' in fl:
-                        iso_file = fl
-                        break
-                logger.debug(iso_file)
-                if not iso_file:
-                    raise Error("Necessary files are missing in hmc")
+            is_img_in_hmc = check_image_in_hmc(module, params)
+            if not is_img_in_hmc:
+                iso_file = image_copy_from_local_to_hmc(module, params)
             else:
-                logger.debug(err)
-                raise Error("could not confirm the necessary image files in hmc")
+                hmc_ls_cmd = "sshpass -p {0} ssh {1}@{2} ls {3}".format(password, hmc_user, hmc_host, params['build_config']['build_file'])
+                rc, out, err = module.run_command(hmc_ls_cmd)
+                if rc == 0:
+                    files = out.split()
+                    iso_count = sum(file.endswith(".iso") for file in files)
+                    if iso_count > 1:
+                        raise Error("There are multiple iso files in specified folder. Please specify the iso file also in build_file")
+                    else:
+                        for fl in files:
+                            if '.iso' in fl:
+                                iso_file = fl
+                                break
+                        if not iso_file:
+                            raise Error("Necessary files are missing in hmc")
+                else:
+                    logger.debug(err)
+                    raise Error("could not confirm the necessary image files in hmc")
 
     otherConfig = {}
     if params['build_config']['hostname']:
@@ -552,9 +586,11 @@ def update_hmc(module, params):
     if locationType == 'disk':
         if not is_img_in_hmc:
             otherConfig['-C'] = ""
-            otherConfig['-F'] = '/home/{0}/network_install/{1}'.format(hmc_user, iso_file)
+            otherConfig['-F'] = '/home/{0}/network_install/{1}'.format(hmc_user, iso_file).strip()
+        elif '.iso' in params['build_config']['build_file']:
+            otherConfig['-F'] = '/{0}'.format(params['build_config']['build_file']).strip()
         else:
-            otherConfig['-F'] = '/{0}/{1}'.format(params['build_config']['build_file'], iso_file)
+            otherConfig['-F'] = '/{0}/{1}'.format(params['build_config']['build_file'], iso_file).strip()
 
     initial_version_details = hmc.listHMCVersion()
 
@@ -563,7 +599,7 @@ def update_hmc(module, params):
         if int(initial_version_details["SERVICEPACK"]) >= 1030:
             otherConfig['--PTF'] = params['build_config']['ptf']
         else:
-            raise VersionError("Update through ibmwebsite supported from 1030 version onwards.")
+            raise VersionError("Update through ibmwebsite supported from V10 R2 M1030 version onwards.")
 
     # this option to restart hmc after configuration
     otherConfig['-R'] = " "